@@ -1,4 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
+using SpiritKing.Controllers;
 
 namespace SpiritKing.Components.Posessables;
 
@@ -6,7 +7,7 @@ internal class Gargoyle : Posessable
 {
     public override Attack NormalAttack { get; set; }
 
-    public Gargoyle(Game game, Vector2 position, GameWorldHandler gameWorld, bool isPosesed = false) : base(game, new Stats()
+    public Gargoyle(Game game, Vector2 position, bool isPosesed = false) : base(game, new Stats()
     {
         WalkSpeed = 175,
         RunSpeed = 320,
@@ -24,9 +25,9 @@ internal class Gargoyle : Posessable
         Health = 100,
         Stamina = 170,
         Color = Color.DarkKhaki
-    }, gameWorld, isPosesed)
+    }, isPosesed)
     {
         Position = position;
-        NormalAttack = new Attack(new CollisionShape(Position.X, Position.Y, 70, 80), 0.2f, 0.2f, 10, 30, 10);
+        NormalAttack = new Attack(Position, new(70, 80), 0.2f, 0.2f, 10, 30, 10);
     }
 }
