@@ -1,4 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
+using SpiritKing.Controllers;
 
 namespace SpiritKing.Components.Posessables;
 
@@ -6,7 +7,7 @@ internal class Hound : Posessable
 {
     public override Attack NormalAttack { get; set; }
 
-    public Hound(Game game, Vector2 position, GameWorldHandler gameWorld, bool isPosesed = false) : base(game, new Stats()
+    public Hound(Game game, Vector2 position, bool isPosesed = false) : base(game, new Stats()
     {
         WalkSpeed = 205,
         RunSpeed = 500,
@@ -24,9 +25,9 @@ internal class Hound : Posessable
         Health = 100,
         Stamina = 210,
         Color = Color.LightSlateGray,
-    }, gameWorld, isPosesed)
+    }, isPosesed)
     {
         Position = position;
-        NormalAttack = new Attack(new CollisionShape(Position.X, Position.Y, 35, 40), 0.2f, 0.2f, 10, 30, 100);
+        NormalAttack = new Attack(Position, new(35, 40), 0.2f, 0.2f, 10, 30, 100);
     }
 }
